feat: 예상정산내역 쿼리 사용 및 에러&로딩페이지 구현
