+++ src/components/Settlements/SettlementsRight/SettlementsExpected/SettlementsUpdate/index.loading.tsx
@@ -3,41 +3,37 @@ import theme from '@styles/theme';
 import Skeleton from 'react-loading-skeleton';
 import 'react-loading-skeleton/dist/skeleton.css';
 
-const SettlementsExpectedLoading = () => {
+const Loading = () => {
   return (
     <Container>
       <Title />
       <ContentsWrapper>
-        <ContentsTop count={2} />
-        <ContentsBottom count={2} />
+        <ContentsTop />
       </ContentsWrapper>
     </Container>
   );
 };
 
-export default SettlementsExpectedLoading;
+export default Loading;
 
 const Container = styled.div`
   width: 100%;
-  height: 100%;
+  height: 155.67px;
 
-  margin-left: 10px;
-  padding: 30px 15px;
-  border-radius: 20px;
+  margin-top: 20px;
+  border-radius: 5px;
 
   display: flex;
   flex-direction: column;
 
   background-color: white;
 
-  ${theme.response.tablet} {
-    width: 90%;
-    max-height: 220px;
+  ${theme.response.tablet} {    
+    border-radius: 5px;
 
-    margin: 0;
-    padding: 10px 10px;
+    gap: 15px;
 
-    overflow: hidden;
+    background-color: #f2f4f5;
   }
 `;
 
@@ -94,19 +90,5 @@ const ContentsTop = styled(BaseSkeleton)`
   ${theme.response.tablet} {
     width: 100%;
     height: 50px;
-  }
-`;
-
-const ContentsBottom = styled(BaseSkeleton)`
-  height: 100px;
-
-  padding: 10px;
-
-  display: flex;
-  flex-direction: column;
-  flex: 1;
-
-  ${theme.response.tablet} {
-    display: none;
   }
 `;
