+++ src/components/Settlements/SettlementsRight/SettlementsExpected/index.tsx
@@ -1,14 +1,17 @@
 import styled from '@emotion/styled';
-import { useRecoilValue } from 'recoil';
-import { useGetSettlemented } from 'src/hooks/queries/useGetSettlemented';
+import { ErrorBoundary } from 'react-error-boundary';
+import { useQueryErrorResetBoundary } from '@tanstack/react-query';
+import { Suspense } from 'react';
 
-import headerAccommodationState from '@recoil/atoms/headerAccommodationState';
 import SyncIcon from '@assets/icons/sync-outline.svg';
 import theme from '@styles/theme';
+import SettlementsUpdate from './SettlementsUpdate';
+import Loading from './SettlementsUpdate/index.loading';
+import ErrorFallback from './SettlementsUpdate/index.error';
 
 const SettlementsExpected = () => {
-  
-  const accommodation = useRecoilValue(headerAccommodationState);
+
+  const { reset } = useQueryErrorResetBoundary();
 
   const currentDate = new Date();
 
@@ -23,8 +26,6 @@ const SettlementsExpected = () => {
   };
 
   const firstDayOfMonth = new Date(currentDate.getFullYear(), currentDate.getMonth(), 1);
-
-  const { data: summary } = useGetSettlemented(accommodation.id);
 
   return (
     <Container>
@@ -54,32 +55,14 @@ const SettlementsExpected = () => {
             {`${formatDate(firstDayOfMonth)} ~ ${formatDate(new Date(currentDate.getFullYear(), currentDate.getMonth() + 1, 0))}`}
             </CouponDay>
           </CouponContainer>
-          <UpdatedInnerContainer>
-            <UpdatedWrapper>
-              <WrapperTop>
-                <DueDateText>
-                  정산예정일
-                </DueDateText>
-                <DueDateDay>
-                {`${currentDate.getFullYear()}년 ${currentDate.getMonth() + 2}월 10일`}
-                </DueDateDay>
-              </WrapperTop>
-              <WrapperBottom>
-                <DueDateText>
-                  {`${currentDate.getMonth() + 1}월 1일부터 현재까지 금액`}
-                </DueDateText>
-                <CommonContainer>
-                  <Icon src={SyncIcon} alt="업데이트" />
-                  <UpdatedText>
-                    매일 00시 00분에 업데이트
-                  </UpdatedText>
-                </CommonContainer>
-                <DueDateDay>
-                {summary ? (summary.this_month_settlement_amount === 0 ? '-' : new Intl.NumberFormat('ko-KR').format(summary.this_month_settlement_amount) + '원') : '데이터 로딩 중...'}
-                </DueDateDay>
-              </WrapperBottom>
-            </UpdatedWrapper>
-          </UpdatedInnerContainer>
+          <ErrorBoundary
+              onReset={reset}
+              fallbackRender={ErrorFallback}
+            >
+          <Suspense fallback={<Loading />}>
+          <SettlementsUpdate />
+          </Suspense>
+          </ErrorBoundary>
         </UpdatedContainer>
       </InnerContainer>
     </Container>
@@ -159,56 +142,6 @@ const UpdatedContainer = styled.div`
 
   border: 1.5px solid white;
   border-radius: 8px;
-`;
-
-const UpdatedInnerContainer = styled.div`
-  margin-top: 20px;
-
-  background-color: white;
-
-  border: 1px solid white;
-  border-radius: 5px;
-`;
-
-const UpdatedWrapper = styled.div`
-  margin: 15px;
-`;
-
-const WrapperTop = styled.div`
-  display: flex;
-  flex-direction: column;
-`;
-
-const DueDateText = styled.div`
-  font-size: 14px;
-  font-weight: bold;
-  color: black;
-
-  @media (max-width: 478px) {
-    font-size: 11px;
-  }
-`;
-
-const DueDateDay = styled.div`
-  margin-left: auto;
-  margin-top: 10px;
-
-  align-items: flex-end
-
-  font-size: 16px;
-  font-weight: bold;
-  color: #3182F6;
-
-  @media (max-width: 478px) {
-    font-size: 10px;
-  }
-`;
-
-const WrapperBottom = styled.div`
-  margin-top: 20px;
-
-  display: flex;
-  flex-direction: column;
 `;
 
 const CouponContainer = styled.div`
