+++ src/components/Settlements/SettlementsRight/index.tsx
@@ -1,44 +1,21 @@
 import styled from '@emotion/styled';
-import { ErrorBoundary } from 'react-error-boundary';
-import { useQueryErrorResetBoundary } from '@tanstack/react-query';
-import { Suspense } from 'react';
 
 import SettlementsBefore from "./SettlementsBefore"
 import SettlementsExpected from "./SettlementsExpected"
 import settlementsLogo from '@assets/icons/settlements-logo.svg'; 
 import theme from '@styles/theme';
-import SettlementsBeforeLoading from './SettlementsBefore/index.loading'
-import SettlementsBeforeErrorFallback from './SettlementsBefore/index.error';
-import SettlementsExpectedLoading from './SettlementsExpected/index.loading'
-import SettlementsExpectedErrorFallback from './SettlementsExpected/index.error';
 
 const SettlementsRight = () => {
-
-  const { reset } = useQueryErrorResetBoundary();
 
   return (
     <Container>
       <InnerContainer>
         <StyledSettlementsExpected>
-          <ErrorBoundary
-              onReset={reset}
-              fallbackRender={SettlementsExpectedErrorFallback}
-            >
-          <Suspense fallback={<SettlementsExpectedLoading />}>
             <SettlementsExpected />
-          </Suspense>
-          </ErrorBoundary>
         </StyledSettlementsExpected>
         <hr />
         <StyledSettlementsBefore>
-          <ErrorBoundary
-            onReset={reset}
-            fallbackRender={SettlementsBeforeErrorFallback}
-          >
-            <Suspense fallback={<SettlementsBeforeLoading />}>
               <SettlementsBefore />
-            </Suspense>
-          </ErrorBoundary>
         </StyledSettlementsBefore>
       </InnerContainer>
       <Logo />
@@ -114,6 +91,7 @@ const StyledSettlementsExpected = styled.div`
     height: 300px;
   }
 `;
+
 const StyledSettlementsBefore = styled.div`
   position: relative;
